requirements.txt
asgiref==3.4.1
Django==4.0
djangorestframework==3.13.0
gunicorn==20.1.0
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
